+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   value) to "Block"
 - Added support for auto generating wiki, help files, markdown linting
   and checking examples.
-- Added CommonResourceHelper.psm1 based on copy from [PSDscResources](https://github.com/PowerShell/PSDscResources/blob/dev/DscResources/CommonResourceHelper.psm1).
+- Added NetworkingDsc.ResourceHelper module based on copy from [PSDscResources](https://github.com/PowerShell/PSDscResources/blob/dev/DscResources/CommonResourceHelper.psm1).
 - MSFT_xFirewall:
   - Cleaned up ParameterList table layout and moved into a new file
     (MSFT_xFirewall.data.psd1).
@@ -18,7 +18,7 @@
     See [GitHub issue](https://github.com/PowerShell/xNetworking/issues/169).
   - Fixed integration tests so that values being set are correctly tested.
   - Added integration tests for Removal of Firewall rule.
-- Added NetworkingCommon module to contain shared networking functions.
+- Added NetworkingDsc.Common module to contain shared networking functions.
 
 ## 3.1.0.0